Fix sorting and refactor
