+++ Source/LibationWinForms/AvaloniaUI/Views/ProductsGrid/ProductsDisplay2.ColumnCustomization.xaml.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace LibationWinForms.AvaloniaUI.Views.ProductsGrid
+{
+	public partial class ProductsDisplay2
+	{
+		private void Configure_ColumnCustomization() { }
+	}
+}
