+++ Source/LibationWinForms/AvaloniaUI/Views/ProductsGrid/ProductsDisplay2.ScanAndRemove.xaml.cs
@@ -1,0 +1,106 @@
+﻿using ApplicationServices;
+using AudibleUtilities;
+using DataLayer;
+using LibationWinForms.AvaloniaUI.ViewModels;
+using ReactiveUI;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace LibationWinForms.AvaloniaUI.Views.ProductsGrid
+{
+	public partial class ProductsDisplay2
+	{
+		private void Configure_ScanAndRemove() { }
+
+		public void CloseRemoveBooksColumn()
+			=> removeGVColumn.IsVisible = false;
+
+		public async Task RemoveCheckedBooksAsync()
+		{
+			var selectedBooks = GetAllBookEntries().Where(lbe => lbe.Remove == true).ToList();
+
+			if (selectedBooks.Count == 0)
+				return;
+
+			var libraryBooks = selectedBooks.Select(rge => rge.LibraryBook).ToList();
+			var result = MessageBoxLib.ShowConfirmationDialog(
+				libraryBooks,
+				$"Are you sure you want to remove {selectedBooks.Count} books from Libation's library?",
+				"Remove books from Libation?");
+
+			if (result != System.Windows.Forms.DialogResult.Yes)
+				return;
+
+			RemoveBooks(selectedBooks);
+			var idsToRemove = libraryBooks.Select(lb => lb.Book.AudibleProductId).ToList();
+			var removeLibraryBooks = await LibraryCommands.RemoveBooksAsync(idsToRemove);
+
+			RemovableCountChanged?.Invoke(this, GetAllBookEntries().Count(lbe => lbe.Remove is true));
+		}
+		public async Task ScanAndRemoveBooksAsync(params Account[] accounts)
+		{
+			RemovableCountChanged?.Invoke(this, 0);
+			removeGVColumn.IsVisible = true;
+
+			try
+			{
+				if (accounts is null || accounts.Length == 0)
+					return;
+
+				var allBooks = GetAllBookEntries();
+
+				foreach (var b in allBooks)
+					b.Remove = false;
+
+				var lib = allBooks
+					.Select(lbe => lbe.LibraryBook)
+					.Where(lb => !lb.Book.HasLiberated());
+
+				var removedBooks = await LibraryCommands.FindInactiveBooks(Login.WinformLoginChoiceEager.ApiExtendedFunc, lib, accounts);
+
+				var removable = allBooks.Where(lbe => removedBooks.Any(rb => rb.Book.AudibleProductId == lbe.AudibleProductId)).ToList();
+
+				foreach (var r in removable)
+					r.Remove = true;
+
+				RemovableCountChanged?.Invoke(this, GetAllBookEntries().Count(lbe => lbe.Remove is true));
+			}
+			catch (Exception ex)
+			{
+				MessageBoxLib.ShowAdminAlert(
+					null,
+					"Error scanning library. You may still manually select books to remove from Libation's library.",
+					"Error scanning library",
+					ex);
+			}
+		}
+
+		private void RemoveBooks(IEnumerable<LibraryBookEntry2> removedBooks)
+		{
+			//Remove books in series from their parents' Children list
+			foreach (var removed in removedBooks.Where(b => b.Parent is not null))
+			{
+				removed.Parent.Children.Remove(removed);
+
+				//In Avalonia, if you fire PropertyChanged with an empty or invalid property name, nothing is updated.
+				//So we must notify for specific properties that we believed changed.
+				removed.Parent.RaisePropertyChanged(nameof(SeriesEntrys2.Length));
+				removed.Parent.RaisePropertyChanged(nameof(SeriesEntrys2.PurchaseDate));
+			}
+
+			//Remove series that have no children
+			var removedSeries =
+				bindingList
+				.AllItems()
+				.EmptySeries();
+
+			foreach (var removed in removedBooks.Cast<GridEntry2>().Concat(removedSeries))
+				//no need to re-filter for removed books
+				bindingList.Remove(removed);
+
+			VisibleCountChanged?.Invoke(this, bindingList.BookEntries().Count());
+		}
+	}
+}
